pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "volatility-relative-value-toolkit"
version = "0.1.0"
description = "Reproducible research toolkit for volatility relative value strategies."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas==2.2.3",
  "numpy==2.1.3",
  "pyarrow==18.1.0",
  "PyYAML==6.0.2",
  "scikit-learn==1.5.2",
  "matplotlib==3.9.2",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "ruff==0.7.4",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
